Don't take DispatchArgs, because we don't need it
This is the key change, because this lets us stop calling
lookupSingletonClass only to have applyTypeArguments immediately call
`attachedClass` again. Now we can pass the generic class directly to
`applyTypeArguments`.

This is critical to unblocking `T.class_of(A)[...]`, because the
singleton class of the singleton class of A probably doesn't exist, but
we don't actually care--we care about just the singleton class of A.