Take `constr` by `const &`
This prevents the class of bug we encounter in the previous commit.

We can afford a deepCopy in this method because this should only be
called when we're sure that we want to report a type error, which are
uncommon and already expensive for all the string allocations they have
to do.