Delete this ENFORCE
While I think this is an admirable ENFORCE, as it's trying to say "don't
allocate if you don't have to," in practice this is really hard to work
with.

The only place that we call `TypeConstraint::deepCopy` is in this PR, so
there are no use cases that currently depend on this ENFORCE. And if we
were to try to change `maybeAutocorrect` to be able to keep this
ENFORCE, it would involve nearly duplicating all the code, or dropping
const-correctness, which defeats the point of the change.