Reduce Sorbet runtime retained memory by releasing location information as soon as possible (#6837)
* Remove unused `loc` parameter from `build_sig` method

* Release declaration block's `loc` as soon as possible

The `loc` field of `DeclarationBlock` is used to store the location of
the signature declaration. However, the value stored in that field is
a `Thread::Backtrace::Location` object that was returned from
`caller_locations`.

Holding onto this object is problematic since `caller_locations`
materializes a `Thread::Backtrace` object, and each
`Thread::Backtrace::Location` object holds a reference to that backtrace.
Before Ruby 3, a call to `caller_locations` would materialize the whole
backtrace, even if the `caller_locations(1, 1)` form was used. Ruby 3
has improved this situation by implementing partial backtraces, but
holding on the the `loc` value still retains a `Location` and `Backtrace`
object regardless.

This means that if Sorbet runtime does not release the location object
as soon as possible, it would end up retaining a lot of backtraces,
namely one for each signature defined. This causes a lot of memory
growth for no good reason. For example, we have a high amount of
"backtrace" objects lying around in Shopify Core monolith memory and
they are marked as `uncollectible`. Peaking at them, the common
relation is:
```
(IMEMO: ment) -> (DATA: proc) -> (IMEMO: env) -> (STRUCT: (null)) -> (DATA: frame_info) -> (DATA: backtrace)
```
which we've found points to the `Kernel.caller_locations(1, 1)&.first`
call inside the `T::Sig#sig` method.

It seems like the location information is only needed to be able to
pass it to `sig_builder_error_handler`, so we should be able to just
release it after we are able to build a `Declaration` object.