Release memory after deduping
We call `sortAndDedupe` to shrink the `nameHashes` field of `UsageHash`.
(This structure is used to remember the names of symbols that a file
might use, so we know which downstream files might need to be
retypechecked.)

This was causing all `FileHash`'s in Stripe's codebase to take up about
6x the space that it needed to. After shrinking, Sorbet uses about 1/6th
the memory to store all `FileHash` objects.

Note that this cost was only paid for cold-cache runs of LSP.
Warm cache runs of LSP would read file hashes from the cache, which
would not have serialized the unused capacity in these vectors.

So while this is definitely an improvement, it's not an improvement that
applies to 100% of LSP initializations. (But it does reduce the worst
case.)