Add editForDSLMethod helper (#6958)
* Use ctx.file instead

* Add editForDSLMethod helper

* Clever use of `""`

* Split Context into GlobalState and FileRef

* Copy over the root logic

This is kind of a tradeoff.

For the SigSuggestion autocorrect, this has the unfortunate side effect
of inserting `extend T::Sig` multiple times in the file. But IMO,
that's kind of find, because that would only apply if there were
multiple methods without a sig in the file, and either

- You used "apply all fixes for file", or
- You used `-a` at the command line

There will be many cases where neither of those is true, which means
that the new behavior is a strict improvement.

In the cases where it does add duplicate, the redundant calls to
`extend T::Sig` are a no-op, so it's not _incorrect_ to have them,
it's just maybe bad style. I think I'd prefer having something that is
really good at fixing type errors.

It's hard to do better here. If it were possible to always insert the
`extend T::Sig` somewhere like "at the top of the file, but after all
requires and comments," then all the autocorrect edits would share the
same location, causing only one of them to get applied (we only apply
the first edit that happens at a given place).

The real goal is to get this in so that other call sites that _are_ fine
with this logic can use this one new helper.

* Also use this helper in calls.cc

* Don't change things too much

* fixup

* stog //test/cli:update

* sbg && tools/scripts/update_testdata_exp.sh --no-build test/testdata/infer/autocorrect_extend_t_sig.rb test/testdata/infer/kwsplat_is_sometimes_okay.rb test/testdata/infer/suggest_useless_sig.rb