Remove Array#+ -> Array#concat autocorrect
This was not great because it suggests people to use a method that
mutates the underlying array when `+` wouldn't do that.

We might want to revive the changes we're removing here and bring them
back suggesting changing `Array#concat` to `Array#+` but people use
`concat` way less as it is, so it's possible that it'll be so rare that
we don't want to justify all this custom code.