Create applied types for type in sealed_subclasses
In recordSealedSubclass, we use `make_type<ClassType>(sym)` instead of
`sym.data(ctx)->externalType(ctx)` because we call recordSealedSubclass
from ResolveCosntantsWalk, before we've created `<AttachedClass>` type
members for all classes.

But that's weird, because all singleton classes are supposed to be
generic classes and are thus supposed to be applied types.

Luckily, there's no way to notice that we're using ClassTypes here, so
we can record the sealed subclasses during ResolveConstantsWalk and then
fix up the type later ResolveTypeMembersAndFieldsWalk (in the portion of
it that runs after we've computed all the type members).

This is a little annoying, because it adds code to the single-threaded
portion of resolver, but I think this is worth it for correctness.