Add a test
Deleting the SelfNew rewriter surfaced this bug on Stripe's codebase.
This bug was actually the pre-existing #3618, not a regression caused by
this change. But deleting the SelfNew rewriter would have made that bug
more common, so I fixed that bug first, and then restacked the SelfNew
PR on that.

I've updated this test to verify that weird combination of two behaviors
doesn't happen. The original failure, after deleting SelfNew and not yet
fixing #3618, was that Sorbet claimed, "Expression does not have a
fully-defined type (Did you reference another class's type members?)"
which was actually "correct"--the attached class type member was not
getting replaced with `Child1` and `Child2` in resultTypeAsSeenFrom.