Widen type of `Array#+` (#7015)
* Widen type of `Array#+`

Historically we held this type artificially narrow because we thought
the error messages would be more direct if we rejected this code. More
context in this change: #4747

But the workarounds are annoying, and the error messages are not
actually that bad. (In particular, the workaround involved mutating the
receiver, something that `Array#+` doesn't do.)

* Add a T.reveal_type

* This autocorrect gets slightly worse

Oh well