This is some weird nonsense
I can't think of a way to fix this right now. I'll have to create a
ticket for it.

The root cause of the bad behavior is actually the changes in 50d2db06b,
it's just that this PR makes the bug happen here too.

I'd really like to not revert the changes in 50d2db06b, because that
supports a feature we know is actually being used, and is closer to the
desired end state. I've spent an afternoon thinking about how to fix the
`Object_class` intrinsic in support of this, but I can't come up with
any simple solutions.