Delete SelfNew (#6978)
* This is a weird example revealing a latent bug

* wip: Remove SelfNew rewriter and intrinsic

* Fix exp file

* Add a test

Deleting the SelfNew rewriter surfaced this bug on Stripe's codebase.
This bug was actually the pre-existing #3618, not a regression caused by
this change. But deleting the SelfNew rewriter would have made that bug
more common, so I fixed that bug first, and then restacked the SelfNew
PR on that.

I've updated this test to verify that weird combination of two behaviors
doesn't happen. The original failure, after deleting SelfNew and not yet
fixing #3618, was that Sorbet claimed, "Expression does not have a
fully-defined type (Did you reference another class's type members?)"
which was actually "correct"--the attached class type member was not
getting replaced with `Child1` and `Child2` in resultTypeAsSeenFrom.

* This is some weird nonsense

I can't think of a way to fix this right now. I'll have to create a
ticket for it.

The root cause of the bad behavior is actually the changes in 50d2db06b,
it's just that this PR makes the bug happen here too.

I'd really like to not revert the changes in 50d2db06b, because that
supports a feature we know is actually being used, and is closer to the
desired end state. I've spent an afternoon thinking about how to fix the
`Object_class` intrinsic in support of this, but I can't come up with
any simple solutions.

* Add another test