Move recordSealedSubclass to ResolveTypeMembersAndFieldsWalk
The idea is that we can use the ancestors stored on a ClassDef tree node
to figure out whether a class was directly a subclass of a sealed class.

I had previously thought we wouldn't be able to do this, becuase I
thought that since we'd have to run in ResolveTypeMembersAndFieldsWalk
that would have meant we'd be after ancestor linearization, which would
have made it hard to see whether a sealed class was a direct parent or a
transitive parent. But by looking at the ancestors node on the ClassDef
we still have that information on hand to disambiguate.

This approach means we can simply remove all the sealed class
registration from ResolveConstantsWalk and defer it to
ResolveTypeMembersAndFieldsWalk, which means we don't have to do an
expensive walk tearing down and building back up large union types--we
can simply use externalType directly in recordSealedSubclass.

The fact that we see ClassType and AppliedType here is subtle, but it's
because `T.class_of(MyModule)` is a ClassType while for something like
`T.class_of(MyClass)`, that would be an applied type (module singleton
classes don't currently have `<AttachedClass>` type members).