Register `class` intrinsic on `Kernel`
This works around a bug where jump to def would not work on calls to
`.class`.

What was happening is that any time we register an intrinsic, if there
is not a method already defined for that intrinsic, we create a new one
on the fly. What usually happens is that later we see an RBI definition
for the intrinsic method too, which fills in things like arity,
location, etc.

But in this case, Object#class doesn't exist in our RBIs, only
Kernel#class does. So there was never a method that would come along and
define things like location information, so jump to definition wouldn't
work.