Refactor logic into specific-purpose components
Issue: Current extension code is difficult to follow as it is hevaily entangled, with componens being created in different places and passed around accessed via callbacks.

Change: the goal of this PR is to move to a world where there are single purpose classes that are easier to maintain (and eventually test standalone).
- `SorbetExtensionContext` is added as a container of references that should be globally accessible.
- Logic specific to track the Sorbet Client status is split from the `SorbetStatusBarEntry` class and moved into a new `SorbetStatusProvider` class.   This new class exposes events to which the `SorbetStatusBarEntry` class can register to. This helped removing significant amount of entanglement and move and simplify the `<start|stop|restart>Sorbet` methods.
- To the extent that it was possible, code was only moved and references updated, without modifying logic in any significant way.