Change how we print T.class_of
If we have type_templates that aren't `<AttachedClass>`, we're going to
show all type_templates, because if the user wants to write what they
see as a type, they're going to have to repeat the arg applied to the
`<AttachedClass>` type_template.

Like before, if we have an arg applied to an `<AttachedClass>`
type_template that differs from the bound, we're going to show that,
even if it's the only thing, because sometimes that's semantically
important.

Otherwise, if the only non-fixed type_template is `<AttachedClass>` and
it's been applied to something that exactly equals its upper bound, we
can skip it.