Fix error with bad call on MetaType
This fakes the dispatch as though it goes to `T::Generic#[]`.
It doesn't actually go there, it goes to `T::Types::ClassOf#[]`, but we
didn't have `T::Type::ClassOf` in our initial empty payload, and I
didn't want to add it just for this method. Realistically, I don't think
anyone ever wants to be able to jump-to-def on the `[]` method in
`T.class_of(...)` syntax so this hack should be fine.