Properly implement `Copy Symbol`  and `Rename Symbol`
`Copy Symbol`  and `Rename Symbol` are bound to its command Ids only after `LanguageClient.onReady` happens. This is incorrect as before commands are bound users would get an error from VSCode saying that the corresponding command IDs are unknown by the extension. It is better to bind the commands early and handle the scenario when the client is not ready.

Refactoring the commands also starts the refcatoring of the `LanguageClient` class (actually, initial refactoring is what showed this issue) and will allow to write tests for each one.