Print associativity for AndType and OrType
When using `TypePtr::toString` to debug problems with union/intersection
types, sometimes the problem comes because two types that look the same
have actually associated differently.

We had some code to print `(...)` when the order of operations mattered,
but we did not have the same code to disambiguate associativity given
the same kind of operation.