Stop exposing `LanguageClient` instance
Stop exposing the inner `LanguageClient` instance from `SorbetLanguageClient` as it leads to confusion.

This PR only encapsulates the former as a private filed and exposes methods being used but that will allow in the future to expose typed, purpose specific methods (e.g. instead of exposing the generic `sendRequest`, we could have better purpose-specific methods like `readFile` and `getSymbolInformation`).