Upgrade mocha
The minitest gem made a change in 5.19 to stop defining a `MiniTest`
namespace:

- <https://github.com/minitest/minitest/blob/master/History.rdoc#label-5.19.0+-2F+2023-07-26>

The correct namespace is `Minitest`.

This exposed a bug in the mocha gem, because it was referencing
`MiniTest` still. That bug was fixed in mocha, but only in version
2.1.0:

- <https://github.com/minitest/minitest/blob/master/History.rdoc#label-5.19.0+-2F+2023-07-26>

Luckily, upgrading to 2.x was not that hard.