Index all `__package.rb` files before everything else
We want to start using the package graph for the sake of skipping some
work, which means we have to make sure that package files are indexed
before we do anything else. Eventually a second step will be to both
index **and** parse the package DB before we do anything else, and I
will do that in a follow up change.

But for the time being, I want to start slowly and make sure nothing is
breaking.

This _should_ be a no-op change, though it will involve one more fork +
join step, which might make Sorbet slower on pay-server because of this.
While that's not great, we should recover that time and more when we
start using this to type check a subset of the codebase.