Change loc for `&&` and `||` condition
The loc for the local variable we used in the `cond` of the `If` that
`&&` and `||` desugar into was previously the whole LHS. In the case of
a long expression on the LHS, that meant that Sorbet would highlight
multiple lines of (irrelevant) code when showing a "This code is
untyped" error.

To my knowledge, the only place in Sorbet where the loc of a local
variable used in an `If` condition would matter is for the error message
about untyped code usages. That is: this loc is not important for any
other autocorrects like adding `T.must` or `T.unsafe`, where the
location of the error is going to get wrapped with a method call.

That means we're free to change this however we like. With that in mind,
I'd like to point to the `... || ...` instead, which makes it a little
more obvious where the "conditional branch on untyped" is coming from.