no-op: Move pipeline::packager out of pipeline::resolve (#7208)
* no-op: Mark EnforcePackagePrefix const

I was curious whether this was actually mutating the tree, and it turns
out it's not. Given the name, I think I'd like to lock in that
property, just so that the first person who breaks it understand when
they've done it.

* fixup: Previous PR forgot to make this edit

See changes in ddf9172f2

* no-op: This code is dead

There is a `return` instruction immediately before.

Stepping back, setPackageNameOnFiles shouldn't need to run, because at
the moment:

- any change to aa package file (even a no-op change) takes the slow
  path
- any change to add a new file or delete a file takes the slow path

So that leaves only pre-existing, non-package files, which means we
shouldn't have to go `setPackageNameOnFiles` on all these files
again--it's still populated correctly.

* no-op: Move pipeline::packager out of pipeline::resolve

This should be a no-op, but sets us up for being able to split apart the
logic in pipeline::packager so that we can populate the package DB
before we index the whole codebase.