Server changes
We have some options here. Another option would be to implement
something like "save all files that weren't open before we applied the
WorkspaceEdit," but that's a little tricky, because we'd have to compute
the list of open files on the server side--by the time the WorkspaceEdit
has applied, VS Code will already have the files open, so we can't do it
purely in the client.

In any case, this isn't a trapdoor decision--we can always decide to
change our minds and expand this feature if people start to ask for it.