Don't register as a command in package.json
This prevents it from being shown in the command picker, but still lets
it be called from the extension.

This reduces the public surface area of the extension. We can always
make this public later if it's something we want. But starting with it
private makes it easier for us to change our minds and either expand or
limit the scope of how this command works.