Auto-save codeAction changes to __package.rb files (#7214)
* no-op: Operate on FileRef more

* no-op: Mark some generated code as const

* Server changes

We have some options here. Another option would be to implement
something like "save all files that weren't open before we applied the
WorkspaceEdit," but that's a little tricky, because we'd have to compute
the list of open files on the server side--by the time the WorkspaceEdit
has applied, VS Code will already have the files open, so we can't do it
purely in the client.

In any case, this isn't a trapdoor decision--we can always decide to
change our minds and expand this feature if people start to ask for it.

* Client changes

* Don't register as a command in package.json

This prevents it from being shown in the command picker, but still lets
it be called from the extension.

This reduces the public surface area of the extension. We can always
make this public later if it's something we want. But starting with it
private makes it easier for us to change our minds and either expand or
limit the scope of how this command works.

* Move log to trace level

* Another warning -> trace log

* sort