Add some ENFORCEs
The `==` check on rubyRegionId should be a stronger claim than the
`==` on `isExceptionHandlingBlock` flags. We should only set
`isExceptionHandlingBlock` if the basic block was created for either a
`rescue` or `ensure`, and both `rescue` and `ensure` basic blocks have
ruby regions that differ from `begin` or `else` (or outside the
`begin`). So it can't happen that you'd have two `rubyRegionId`s that
were equal but corresponded to blocks on opposite sides of the exception
handling.