Fix a typo
I've thought about this for the better part of the day to see if I could
craft a test case where it matters whether you wrote thenb or elseb
there, but I've basically come to the conclusion that the ways you could
satisfy conditional where `... == thenb->rubyRegionId` but
`... != elseb->rubyRegionId` are not actually possible because of the
way that we set up rescue/ensure/else blocks.

So I'm just going to change this and not write a test for it because it
should matter (spiritually its a typo) but it doesn't seem to matter
(can't craft a test case where it actually matters).