Add an autocorrect for missing abstract method definitions. (#7048)
* Move LSP's method pretty printer to shared lib

* Support concrete method output in pretty printer

* Add autocorrect for missing abstract method def

* Remove unused import

* Move source gen helpers to their own package

* Take ShowOptions in prettyTypeForMethod

This also fixes LSP hover behaviour because the appropriate options are
now passed through to the methods that generate the text that appears on
hover.

* Rename showForRBI to useValidSyntax

* Correct default argument suggestions

* Convert to autocorrect tests

* Fix formatting

* Fix tools BUILD file

* Comment on why we intentionally introduce another error

* Remove stray StrCat

* Remove some unnecessary includes

* Prefer shorter name in autocorrect text

* Insert autocorrect code at bottom of class

Also add a test that demonstrates that the new autocorrect properly
handles declarations that use a class/module constructor.

* Roll all method definitions into single fix

As opposed to a separate fix for each method. This makes it easier to
define several new methods correctly.

This commit also adds support for autocorrecting missing method
definitions on the most common type of stub class/module definition:
`class Foo < Bar; end`.

* Fix LSP tests

* Sort missing methods in autocorrect output

* Pass strings by reference

* Use early returns more consistently

* Deduplicate corrections for the same method

* Correct method show options used in LSP hover

* Correct error message in test

* Prefer passing around core::Context explicitly

Rather than allowing a value of that type to be implicitly converted to
a GlobalState reference. This lets us be clearer about computing certain
locations relative to the enclosing context.

* Inspect AST to determine if class body is empty

This is more rigorous than the previous regex-based logic.

* Reinstate test that demonstrates hover display options

Invalid syntax is allowed on hover.