Only use `initialize` DispatchResult if we found something
It's a lot less error prone to take the whole dispatch component, rather
than pluck individual pieces off of it. But we should only do that if we
actually _want_ it. If the dispatch to `initialize` from `Class#new`
fails, we would overwrite the original dispatch to `Class#new`, and then
claw back some of the things we overwrote. While working on typing
super, I realized that we weren't copying the `constr` back if
`initialize` didn't exist, so it could end up that you'd dispatch to a
method that `.exists()` but you'd have `constr == nullptr` which is bad.

We had a test for this on the WIP changes to super. I wanted to do this
refactor as pre-work to prove that it's a no-op normally. This will end
up getting tested indirectly by way of the tests that will come in
service of `super`.