Stop special-casing `initialize`
The previous commit made it so that if we didn't successfully dispatch
to an `initialize` method, we don't report any errors.

But we actually do want some errors: if you dispatch to `initialize` and
it "didn't exist" that used to mean that it dispatched to
`BasicObject#initialize`. There was some comment about why that was
undesirable to model in RBIs, but I think that doesn't apply anymore. We
can simply let the dispatch to `BasicObject#initialize` happen, and then
Sorbet will report errors like normal because `BasicObject#initialize`
does not take any arguments.