Move super handling to desugar
local_vars runs after we've already flattened methods, which means we
don't remember whether a method was inside a block or not during the
local_vars walk.

This is slightly annoying because it means that we don't get to defer
the super handling until after the rewriter passes. But it lets us more
accurately tell whether we're inside a method that's inside a block
(which might be getting passed to `class_eval` or some other DSL method)