Type check calls to `super` in classes (#7231)
* Only use `initialize` DispatchResult if we found something

It's a lot less error prone to take the whole dispatch component, rather
than pluck individual pieces off of it. But we should only do that if we
actually _want_ it. If the dispatch to `initialize` from `Class#new`
fails, we would overwrite the original dispatch to `Class#new`, and then
claw back some of the things we overwrote. While working on typing
super, I realized that we weren't copying the `constr` back if
`initialize` didn't exist, so it could end up that you'd dispatch to a
method that `.exists()` but you'd have `constr == nullptr` which is bad.

We had a test for this on the WIP changes to super. I wanted to do this
refactor as pre-work to prove that it's a no-op normally. This will end
up getting tested indirectly by way of the tests that will come in
service of `super`.

* Stop special-casing `initialize`

The previous commit made it so that if we didn't successfully dispatch
to an `initialize` method, we don't report any errors.

But we actually do want some errors: if you dispatch to `initialize` and
it "didn't exist" that used to mean that it dispatched to
`BasicObject#initialize`. There was some comment about why that was
undesirable to model in RBIs, but I think that doesn't apply anymore. We
can simply let the dispatch to `BasicObject#initialize` happen, and then
Sorbet will report errors like normal because `BasicObject#initialize`
does not take any arguments.

* Update some tests

* Preserve error messages

* Add BasicObject_initialize

* Add a test and update others

* add new test

* keep the original loc of the arg in local_vars

* split off findMemberTransitiveAncestors

* add enclosingMethodForSuper to dispatchArgs

* untyped super for super calls in modules and blocks

* don't forward synthetic block args in typed: strict

* handle super

* update sigs

* update error messages

* add a TODO

* remove this file

* format

* get rid of remaining noName()s

* update tests

* strict version of test/testdata/local_vars/z_super_args

* Show test behavior before

* wip

* Hide the `dealias` and `depth` parameters

* Only do one call to findMemberTransitive for super

* Rename to findParentMethodTransitive

* Add a new error code for this, just in case

* Factor out targetName variable

* Add some tests for the things I just fixed

* Also add docs to website

* Explicitly ignore `<super>` for did you mean/autocorrects

We will have to circle back on these error messages.

* Fix trailing comma

* Landed upstream

* Revert completion test changes

* Stop setting the loc to the parameter list

* Use a different way to detect when we're passing a fake block arg

* Whoops, need the end loc

* Add a command line option to toggle typed super

* Add docs to FAQ for `super`

* wip

* Fix some autocorrects

* Last minute fixups

* Some new tests

* Add this error note

* Add comment

* stog //test/cli:update

* Fix compiler

* Properly fix zsuper in typed: true

* Add support for typed-super=false to tests

* Fix tests

* Fix one more test

* Go back to what it was

* Use targetName in one more place

* wip

* Fix typo

* Add Nathan's test cases, fixes

* Whoops

* Move super handling to desugar

local_vars runs after we've already flattened methods, which means we
don't remember whether a method was inside a block or not during the
local_vars walk.

This is slightly annoying because it means that we don't get to defer
the super handling until after the rewriter passes. But it lets us more
accurately tell whether we're inside a method that's inside a block
(which might be getting passed to `class_eval` or some other DSL method)

* Fix tests

* Move comment

* Remove more stuff from local_vars

* Don't suggest `--typed-super=false` in error message

* Fix tests

---------

Co-authored-by: Neil Parikh <parikh.neil@me.com>