Misuse of `static` no longer falsifies untyped blame
The first time these `arrayOfUntyped` / `rangeOfUntyped` /
`hashOfUntyped` methods were called, the blame was being cached in a
`static` variable and then reused throughout the lifetime of the
program.

In Stripe's codebase, we were seeing a lot of untyped blame to
`<build-hash>` and none to `<shapeUnderlying>`, which I thought was
strange because a program as simple as:

    res = {foo: 0}
    foo = res[:foo].even?

should be enough to show an untyped blame to `<shapeUnderlying>`... and
on small examples it was!

I only realized that something was up because I tried adding some extra
error sections to the "usage of untyped" error, and I noticed that the
blame symbol was different in my editor vs at the command line on small
examples. This was enough to let me run the code under a debugger which
eventually led me to these lines.