Switch to checking procs.rbi into the repo
As of Bazel 6 the `bazel-out/host/` path is no longer stable:

> genrule switched to use exec transition instead of host. This can
> break targets with hardcoded output paths. To avoid using hardcoded
> paths use make variables, see
> https://docs.bazel.build/versions/4.2.2/be/make-variables.html#predefined_variables

The suggestion of using "make variables" doesn't work for us, because
the hardcoded path is inferred from the file off the disk, instead of
being a path provided in some rule attribute.

Since this file changes infrequently, I've opted to make this file
checked into the repo. I've done that by way of a test that asserts the
generated contents match the contents in the repo, and an executable
that copies the generated file into the repo.

I've also made the suffix of the generated RBI file not have a `.rbi`
file extension, so that it is not picked up by our filegroup glob for
all RBI files, meaning that the `procs.rbi` file only shows up in the
payload once.