Update this test
The --censor-for-snapshot-tests flag explicitly only prints the first 10
files in this test, so that adding a new RBI file doesn't involve making
a change to this test, which trips up new contributors.

I double checked that without that flag, the procs.rbi file shows up
here, and at the path we expect:

    +  {
    +   "path": "https://github.com/sorbet/sorbet/tree/master/rbi/procs.rbi",
    +   "sigil": "Strict",
    +   "strict": "Strict",
    +   "min_error_level": "Max"
    +  },

That being said, it no longer shows up in the test snapshot because it
sorts alphabetically lower. I'm fine with that.