Always show the sectionHeader in LSP diagnostics (#7566)
* Remove `\n` character from message

In languages that used markdown to render the message to HTML, this
would just get treated as useless leading whitespace and ignored. And in
other clients, this would get rendered as an actual newline, but it
would just look out of place.

* Flip location logic to top of method

If we're going to early exit, no need to do a bunch of string
formatting.

* Always show the sectionHeader in LSP diagnostics

There's frequently important message in the LSP section header.
Sometimes it's just "Autocorrect available" or "Note" or something dumb,
but sometimes it's important. In any case, if it's good enough for the
command line output, it should be good enough for LSP. We shouldn't
leave out critical information.