Use .eql? instead of ==
The default implementation of `==` on `T::Types::Base` uses string
comparison to figure out whether two types are equal (which is a hack
but it's what we have).

Based on how the implementation works, we might end up calling methods
which have been overridden, like `.name` or `.to_s`, which cause code to
run in an unexpected order. In Stripe's codebase, we had example where
the `to_s` functions were expecting certain instance variables to have
already been populated.