Fix uncaught metrics API errors (#7590)
This was showing up in development. I don't know why the
`sorbet.metrics.getExportedApi` command is defined but returning
something with an interface that doesn't match what's expected. This
doesn't appear to happen in Stripe's codebase so we're probably setting
things up correctly there.

It looks like the intention was to gracefully fail for this anyways, so
let's gracefully fail.