Stop showing definition results for literals
Other plugins typically try to associate definitions with things inside
string literals, e.g. by treating them as relative paths or something
like that.

Having Sorbet treat the "definition" of a string literal as the location
of the literal is not useful, and conflicts with other plugins that try
to provide their own definitions for string literals.